Refactoring graph code
